// web/app/components/workflow/types.ts
export const BlockEnum = {
  Start: 'start',
  StartPlaceholder: 'start-placeholder',
  End: 'end',
  Answer: 'answer',
  LLM: 'llm',
  KnowledgeRetrieval: 'knowledge-retrieval',
  QuestionClassifier: 'question-classifier',
  IfElse: 'if-else',
  Code: 'code',
  TemplateTransform: 'template-transform',
  HttpRequest: 'http-request',
  VariableAssigner: 'variable-assigner',
  VariableAggregator: 'variable-aggregator',
  Tool: 'tool',
  ParameterExtractor: 'parameter-extractor',
  Iteration: 'iteration',
  DocExtractor: 'document-extractor',
  ListFilter: 'list-operator',
  IterationStart: 'iteration-start',
  Assigner: 'assigner', // is now named as VariableAssigner
  Agent: 'agent',
  AgentV2: 'agent-v2',
  IntentExecutor: 'intent-executor',
  Loop: 'loop',
  LoopStart: 'loop-start',
  LoopEnd: 'loop-end',
  HumanInput: 'human-input',
  DataSource: 'datasource',
  DataSourceEmpty: 'datasource-empty',
  KnowledgeBase: 'knowledge-index',
  TriggerSchedule: 'trigger-schedule',
  TriggerWebhook: 'trigger-webhook',
  TriggerPlugin: 'trigger-plugin',
} as const

export type BlockEnum = (typeof BlockEnum)[keyof typeof BlockEnum]
// ...
export type Branch = {
  id: string
  name: string
}

export type CommonNodeType<T = {}> = {
  _connectedSourceHandleIds?: string[]
  _connectedTargetHandleIds?: string[]
  _targetBranches?: Branch[]
  _isSingleRun?: boolean
  _runningStatus?: NodeRunningStatus
  _runningBranchId?: string
  _singleRunningStatus?: NodeRunningStatus
  _isCandidate?: boolean
  _isBundled?: boolean
  _children?: { nodeId: string; nodeType: BlockEnum }[]
  _isEntering?: boolean
  _showAddVariablePopup?: boolean
  _holdAddVariablePopup?: boolean
  _iterationLength?: number
  _iterationIndex?: number
  _waitingRun?: boolean
  _retryIndex?: number
  _dataSourceStartToAdd?: boolean
  _isTempNode?: boolean
  _openInlineAgentPanel?: boolean
  isInIteration?: boolean
  iteration_id?: string
  selected?: boolean
  title: string
  desc: string
  type: BlockEnum
  width?: number
  height?: number
  position?: XYPosition
  _loopLength?: number
  _loopIndex?: number
  isInLoop?: boolean
  loop_id?: string
  error_strategy?: ErrorHandleTypeEnum
  retry_config?: WorkflowRetryConfig
  default_value?: DefaultValueForm[]
  credential_id?: string
  subscription_id?: string
  provider_id?: string
  _dimmed?: boolean
} & T &
  Partial<
    Pick<
      PluginDefaultValue,
      'provider_type' | 'provider_name' | 'plugin_id' | 'plugin_unique_identifier'
    >
  >
// ...
export type Node<T = {}> = ReactFlowNode<CommonNodeType<T>>
// ...
export const NodeRunningStatus = {
  NotStart: 'not-start',
  Waiting: 'waiting',
  Listening: 'listening',
  Running: 'running',
  Succeeded: 'succeeded',
  Failed: 'failed',
  Exception: 'exception',
  Retry: 'retry',
  Stopped: 'stopped',
  Paused: 'paused',
} as const

export type NodeRunningStatus = (typeof NodeRunningStatus)[keyof typeof NodeRunningStatus]